Extracted components from StreamSessionsTable
