+++ packages/www/components/Dashboard/StreamSessionsTable/index.tsx
@@ -1,92 +1,14 @@
 import { useCallback, useMemo } from "react";
 import { useApi } from "../../../hooks";
 import Table, { Fetcher, useTableState } from "components/Dashboard/Table";
-import { FilterItem, formatFiltersForApiRequest } from "../Table/filters";
-import TextCell, { TextCellProps } from "components/Dashboard/Table/cells/text";
-import DateCell, { DateCellProps } from "components/Dashboard/Table/cells/date";
-import DurationCell, {
-  DurationCellProps,
-} from "components/Dashboard/Table/cells/duration";
+import { Heading, useSnackbar } from "@livepeer/design-system";
 import {
-  dateSort,
-  numberSort,
-  stringSort,
-} from "components/Dashboard/Table/sorts";
-import Link from "next/link";
-import { SortTypeArgs } from "components/Dashboard/Table/types";
-import {
-  CellComponentProps,
-  TableData,
-} from "components/Dashboard/Table/types";
-import { truncate } from "../../../lib/utils";
-import {
-  Box,
-  Flex,
-  Heading,
-  Link as A,
-  Text,
-  HoverCardRoot,
-  HoverCardContent,
-  HoverCardTrigger,
-  useSnackbar,
-} from "@livepeer/design-system";
-import { ArrowRightIcon, CopyIcon, DownloadIcon } from "@radix-ui/react-icons";
-import { CopyToClipboard } from "react-copy-to-clipboard";
-
-function makeMP4Url(hlsUrl: string, profileName: string): string {
-  const pp = hlsUrl.split("/");
-  pp.pop();
-  return pp.join("/") + "/" + profileName + ".mp4";
-}
-
-type Profile = { name: string; width: number; height: number };
-export type RecordingUrlCellProps = {
-  children?: React.ReactNode;
-  tooltipChildren?: React.ReactNode;
-  mp4Url?: string;
-  id?: string;
-  profiles?: Array<Profile>;
-  showMP4: boolean;
-};
-
-const RecordingUrlCell = <D extends TableData>({
-  cell,
-}: CellComponentProps<D, RecordingUrlCellProps>) => {
-  const id = cell.value.id;
-  return (
-    <Box id={`mp4-link-dropdown-${id}`} css={{ position: "relative" }}>
-      {cell.value.mp4Url ? (
-        <Flex css={{ ai: "center", justifyContent: "space-between" }}>
-          {cell.value.children}
-          {cell.value.showMP4 && cell.value.profiles?.length ? (
-            <A target="_blank" href={makeMP4Url(cell.value.mp4Url, "source")}>
-              <DownloadIcon />
-            </A>
-          ) : null}
-        </Flex>
-      ) : (
-        truncate(cell.value.children, 20)
-      )}
-    </Box>
-  );
-};
-
-const filterItems: FilterItem[] = [
-  { label: "Created Date", id: "createdAt", type: "date" },
-  {
-    label: "Duration (in minutes)",
-    id: "sourceSegmentsDuration",
-    type: "number",
-  },
-];
-
-type SessionsTableData = {
-  id: string;
-  parentStream: TextCellProps;
-  recordingUrl: TextCellProps;
-  createdAt: DateCellProps;
-  sourceSegmentsDuration: DurationCellProps;
-};
+  filterItems,
+  makeColumns,
+  rowsPageFromState,
+  SessionsTableData,
+} from "./helpers";
+import EmptyState from "./EmptyState";
 
 const StreamSessionsTable = ({ title = "Sessions" }: { title?: string }) => {
   const { user, getStreamSessionsByUserId } = useApi();
@@ -94,190 +16,34 @@ const StreamSessionsTable = ({ title = "Sessions" }: { title?: string }) => {
     tableId: "allSessionsTable",
   });
   const [openSnackbar] = useSnackbar();
-
-  const columns = useMemo(
-    () => [
-      {
-        Header: "Stream",
-        accessor: "parentStream",
-        Cell: TextCell,
-        sortType: (...params: SortTypeArgs) =>
-          stringSort("original.parentStream.name", ...params),
-      },
-      {
-        Header: "Created at",
-        accessor: "createdAt",
-        Cell: DateCell,
-        sortType: (...params: SortTypeArgs) =>
-          dateSort("original.createdAt.date", ...params),
-      },
-      {
-        Header: "Duration",
-        accessor: "sourceSegmentsDuration",
-        Cell: DurationCell,
-        sortType: (...params: SortTypeArgs) =>
-          numberSort(
-            "original.sourceSegmentsDuration.sourceSegmentsDuration",
-            ...params
-          ),
-      },
-      {
-        Header: "Recording URL",
-        accessor: "recordingUrl",
-        Cell: RecordingUrlCell,
-        disableSortBy: true,
-      },
-    ],
-    []
-  );
+  const columns = useMemo(makeColumns, []);
 
   const fetcher: Fetcher<SessionsTableData> = useCallback(
-    async (state) => {
-      const [streams, nextCursor, count] = await getStreamSessionsByUserId(
+    async (state) =>
+      rowsPageFromState(
+        state,
         user.id,
-        state.cursor,
-        state.pageSize,
-        state.order,
-        formatFiltersForApiRequest(state.filters, {
-          parseNumber: (n) => n * 60,
-        }),
-        true
-      );
-
-      return {
-        nextCursor,
-        count,
-        rows: streams.map((stream: any) => {
-          return {
-            id: stream.id,
-            parentStream: {
-              id: stream.parentId,
-              name: stream.parentStream.name,
-              children: (
-                <A variant="primary" as={Box}>
-                  {stream.parentStream.name}
-                </A>
-              ),
-              tooltipChildren: stream.createdByTokenName ? (
-                <>
-                  Created by stream <b>{stream.parentStream.name}</b>
-                </>
-              ) : null,
-              href: `/dashboard/streams/${stream.parentId}`,
-            },
-            recordingUrl: {
-              id: stream.id,
-              showMP4: true,
-              profiles:
-                stream.recordingUrl &&
-                stream.recordingStatus === "ready" &&
-                stream.profiles?.length
-                  ? [{ name: "source" }, ...stream.profiles]
-                  : undefined,
-              children:
-                stream.recordingUrl && stream.recordingStatus === "ready" ? (
-                  <HoverCardRoot openDelay={200}>
-                    <HoverCardTrigger>
-                      <Flex css={{ ai: "center" }}>
-                        <CopyToClipboard
-                          text={stream.recordingUrl}
-                          onCopy={() => openSnackbar("Copied to clipboard")}>
-                          <Flex
-                            css={{
-                              cursor: "pointer",
-                              fontSize: "$1",
-                              ai: "center",
-                            }}>
-                            <Box css={{ mr: "$1" }}>
-                              {truncate(stream.recordingUrl, 24)}
-                            </Box>
-                            <CopyIcon />
-                          </Flex>
-                        </CopyToClipboard>
-                      </Flex>
-                    </HoverCardTrigger>
-                    <HoverCardContent>
-                      <Text
-                        variant="gray"
-                        css={{
-                          backgroundColor: "$panel",
-                          borderRadius: 6,
-                          px: "$3",
-                          py: "$1",
-                          fontSize: "$1",
-                          display: "flex",
-                          ai: "center",
-                        }}>
-                        <Box css={{ ml: "$2" }}>{stream.recordingUrl}</Box>
-                      </Text>
-                    </HoverCardContent>
-                  </HoverCardRoot>
-                ) : (
-                  <Box>—</Box>
-                ),
-              mp4Url: stream.recordingUrl ? stream.recordingUrl : undefined,
-            },
-            sourceSegmentsDuration: {
-              sourceSegmentsDuration: stream.sourceSegmentsDuration || 0,
-              status: stream.recordingStatus,
-            },
-            createdAt: {
-              date: new Date(stream.createdAt),
-              fallback: <i>unseen</i>,
-            },
-          };
-        }),
-      };
-    },
+        getStreamSessionsByUserId,
+        openSnackbar
+      ),
     [getStreamSessionsByUserId, user.id]
   );
 
-  const emptyState = (
-    <Flex
-      direction="column"
-      justify="center"
-      css={{
-        margin: "0 auto",
-        height: "calc(100vh - 400px)",
-        maxWidth: 450,
-      }}>
-      <Heading css={{ fontWeight: 500, mb: "$3" }}>No sessions</Heading>
-      <Text variant="gray" css={{ lineHeight: 1.5, mb: "$3" }}>
-        Stream sessions belong to parent streams.
-      </Text>
-      <Link
-        href="https://docs.livepeer.studio/reference/api/get-session"
-        passHref>
-        <A
-          target="_blank"
-          variant="primary"
-          css={{ display: "flex", ai: "center", mb: "$5" }}>
-          <Box>Learn more</Box>
-          <ArrowRightIcon />
-        </A>
-      </Link>
-    </Flex>
-  );
-
   return (
-    <>
-      <Table
-        {...tableProps}
-        columns={columns}
-        fetcher={fetcher}
-        initialSortBy={[{ id: "createdAt", desc: true }]}
-        showOverflow={true}
-        filterItems={filterItems}
-        emptyState={emptyState}
-        header={
-          <>
-            <Heading size="2" css={{ fontWeight: 600 }}>
-              {title}
-            </Heading>
-          </>
-        }
-      />
-    </>
+    <Table
+      {...tableProps}
+      columns={columns}
+      fetcher={fetcher}
+      initialSortBy={[{ id: "createdAt", desc: true }]}
+      showOverflow={true}
+      filterItems={filterItems}
+      emptyState={<EmptyState />}
+      header={
+        <Heading size="2" css={{ fontWeight: 600 }}>
+          {title}
+        </Heading>
+      }
+    />
   );
 };
